soupault: don't ignore drafts
We remove "draft" from the `ignore_extensions` list so that Soupault
will process drafts.

However, we don't want drafts to be checked into Git, so we need to make
sure the build directory has a .gitignore which excludes drafts.

However, we don't want this .gitignore to itself be checked in (it's not
part of the site), so it needs to exclude itself.

However, this means we can't have the .gitignore be simply copied from
the root directory, since we don't want the root .gitignore to ignore
itself.

So, we create a new file, `.build-gitignore`, for these rules, and tell
Soupault that it should be renamed to `.gitignore` in build/ using a
custom asset processor.