minefield: change starting safety logic
The game needs to ensure that the first uncover reveals a 0. Previously
this was accomplished by moving east from the origin by 3 steps each
time until we found a 0. This commit changes it to artificially set the
neighborhood around the origin to be clear of mines. This should be
basically the same -- since each choice of mine is made independently
from all others, retrying until we get a particular condition is
identical to artificially making that condition true. The only
difference is that previously you could assume that the cells directly
west of the starting position had a slightly higher density of mines
than average, which barely useful information and not a good feature.)